+++ refbox-embark.el
@@ -109,14 +109,6 @@
     (refbox-embark--define-action
      map "i" "insert or edit" #'refbox-embark-insert-edit)
     (refbox-embark--define-action
-     map "a" "add file to library" #'refbox-embark-add-file)
-    (refbox-embark--define-action
-     map "A" "attach library files" #'refbox-embark-attach-file)
-    (refbox-embark--define-action
-     map "b" "insert bibtex entry" #'refbox-embark-insert-bibtex)
-    (refbox-embark--define-action
-     map "B" "insert raw bibtex entry" #'refbox-embark-insert-raw-entry)
-    (refbox-embark--define-action
      map "e" "open bibtex entry" #'refbox-embark-open-entry)
     (refbox-embark--define-action
      map "f" "open library files" #'refbox-embark-open-files)
@@ -128,10 +120,6 @@
      map "o" "open resources" #'refbox-embark-open)
     (refbox-embark--define-action
      map "r" "copy reference" #'refbox-embark-copy-reference)
-    (refbox-embark--define-action
-     map "s" "open bibliography source" #'refbox-embark-open-source)
-    (refbox-embark--define-action
-     map "C" "copy references" #'refbox-embark-copy-references)
     (refbox-embark--define-action
      map "RET" "run default action" #'refbox-embark-run-default-action)
     map)